refactor: ContextDSAModule + GenerationDSAModule own perf data (AIC-538)
Phase 3.4 of the SDK refactor. The two DSA (DeepSeek Sparse Attention)
module-level ops move out of ``_legacy.py`` into ``operations/dsa.py``.
Same migration pattern as previous Phase 2/3 commits.

What lives in ``operations/dsa.py`` now:
- ``ContextDSAModule.load_data`` — parses context_dsa_module CSV,
  deepcopies the raw rows BEFORE extrapolation runs (needed by the
  topk-boundary piecewise interp from PR #903), applies grid
  extrapolation, binds both ``database._context_dsa_module_data`` and
  ``database._raw_context_dsa_module_data``.
- ``GenerationDSAModule.load_data`` — parses generation_dsa_module CSV,
  applies grid extrapolation, binds
  ``database._generation_dsa_module_data``. No raw cache (generation
  path uses straight 3D cubic, no topk-piecewise lookup).
- ``_query_context_dsa_module_table`` / ``_query_generation_dsa_module_table``
  classmethods — verbatim ports of the legacy ``PerfDatabase.query_*``
  bodies including the full SOL FLOP/byte accounting, the prefix-
  correction multiplier in context, and the indexer-skip fast path
  when ``full_s <= index_topk``.
- ``_is_dsa_interpolation_miss`` / ``_format_dsa_unavailable_message``
  module-level helpers — pulled from ``PerfDatabase`` since they're
  used only by the two DSA query paths.

What changes in ``perf_database.py``:
- ``__init__`` no longer eagerly loads ``_context_dsa_module_data`` /
  ``_generation_dsa_module_data`` /
  ``_raw_context_dsa_module_data``. The legacy ``__init__`` accidentally
  loaded DSA twice (once near the MLA/Mamba block, once after the
  TensorRT-LLM WideEP block); both loads are gone, replaced by single
  ``ContextDSAModule.load_data(self)`` + ``GenerationDSAModule.load_data(self)``
  calls.
- Two ~70-line DSA extrapolation blocks removed from ``__init__``.
- ``query_context_dsa_module`` / ``query_generation_dsa_module`` reduced
  to one-line delegations.
- ``_interp_dsa_context_topk_piecewise_from_raw`` thin wrapper stays
  because tests at ``test_dsa_module.py:78`` exercise it directly.
- ``_is_dsa_interpolation_miss`` / ``_format_dsa_unavailable_message``
  helpers retained as dead code in ``perf_database.py`` (no remaining
  callers); the cleanup PR prunes them.

Cache key matches every other migrated op:
``(systems_root, system, backend, version, enable_shared_layer)``. The
DSA-specific data structures (``data[fmha_mode][kv_dtype][gemm_mode]
[arch][num_heads][s][b]`` for context, ``data[kv_dtype][gemm_mode]
[arch][num_heads][b][s]`` for generation) match the legacy layout.

Behavior preservation:
- PR #903 topk-boundary regime-aware piecewise interpolation: preserved
  via ``interpolation.interp_dsa_context_topk_piecewise_from_raw``
  call inside ``_query_context_dsa_module_table``.
- PR #956 source-tag attribution: preserved (silicon for table hits,
  sol/empirical for formula paths).
- Prefix-correction multiplier in context (``correction = target_sol /
  base_sol`` when prefix > 0): preserved.
- Indexer-skip fast path when full_s <= index_topk: preserved
  (FLOPs/bytes set to 0).
- HYBRID-mode fallback to empirical, ``PerfDataNotAvailableError`` paths,
  ``_is_dsa_interpolation_miss`` error matching: all preserved.

Verification:
- 670 unit tests pass + 1 skipped.
- 38 regression snapshots bit-identical + 6 skipped (same as Phase 3.1 baseline).
- ruff check + format clean.
