+++ src/aiconfigurator/sdk/operations/_legacy.py
@@ -1199,120 +1199,6 @@ class Mamba2(Operation):
         return self._weights * self._scale_factor
 
 
-# ═══════════════════════════════════════════════════════════════════════
-# DSA (DeepSeek Sparse Attention) Operations
-# ═══════════════════════════════════════════════════════════════════════
-
-
-class ContextDSAModule(Operation):
-    """
-    Context phase DSA (DeepSeek Sparse Attention) module-level operation.
-
-    Models the full DSA attention block including:
-    - kv_a_proj_with_mqa GEMM (includes indexer K projection)
-    - LayerNorm + q_b_proj GEMM
-    - Indexer: wq_b GEMM, weights_proj GEMM, FP8 MQA logits, TopK selection
-    - Sparse MLA attention (attends to top-k tokens instead of full sequence)
-    - BMM pre/post (weight absorption + V projection)
-    - o_proj GEMM
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        num_heads: int,
-        kvcache_quant_mode: common.KVCacheQuantMode,
-        fmha_quant_mode: common.FMHAQuantMode,
-        gemm_quant_mode: common.GEMMQuantMode,
-        architecture: str = "DeepseekV32ForCausalLM",
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._num_heads = num_heads
-        self._kvcache_quant_mode = kvcache_quant_mode
-        self._fmha_quant_mode = fmha_quant_mode
-        self._gemm_quant_mode = gemm_quant_mode
-        self._architecture = architecture
-        self._weights = 0.0
-
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        """Query context DSA latency with energy data."""
-        batch_size = kwargs.get("batch_size")
-        isl = kwargs.get("s")
-        prefix = kwargs.get("prefix", 0)
-
-        result = database.query_context_dsa_module(
-            b=batch_size,
-            s=isl,
-            prefix=prefix,
-            num_heads=self._num_heads,
-            kvcache_quant_mode=self._kvcache_quant_mode,
-            fmha_quant_mode=self._fmha_quant_mode,
-            gemm_quant_mode=self._gemm_quant_mode,
-            architecture=self._architecture,
-        )
-        return PerformanceResult(
-            float(result) * self._scale_factor,
-            energy=result.energy * self._scale_factor,
-            source=getattr(result, "source", "silicon"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-
-class GenerationDSAModule(Operation):
-    """
-    Generation phase DSA (DeepSeek Sparse Attention) module-level operation.
-
-    Models the full DSA attention block during decode:
-    - Same components as ContextDSAModule
-    - Uses paged MQA logits for indexer
-    - Sparse MLA with KV cache lookup
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        num_heads: int,
-        kv_cache_dtype: common.KVCacheQuantMode,
-        gemm_quant_mode: common.GEMMQuantMode,
-        architecture: str = "DeepseekV32ForCausalLM",
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._num_heads = num_heads
-        self._kv_cache_dtype = kv_cache_dtype
-        self._gemm_quant_mode = gemm_quant_mode
-        self._architecture = architecture
-        self._weights = 0.0
-
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        """Query generation DSA latency with energy data."""
-        beam_width = kwargs.get("beam_width")
-        if beam_width != 1:
-            raise ValueError(f"{self.__class__.__name__} only supports beam_width=1, got {beam_width}")
-        batch_size = kwargs.get("batch_size")
-        s = kwargs.get("s")
-
-        result = database.query_generation_dsa_module(
-            b=batch_size,
-            s=s,
-            num_heads=self._num_heads,
-            kv_cache_dtype=self._kv_cache_dtype,
-            gemm_quant_mode=self._gemm_quant_mode,
-            architecture=self._architecture,
-        )
-        return PerformanceResult(
-            float(result) * self._scale_factor,
-            energy=result.energy * self._scale_factor,
-            source=getattr(result, "source", "silicon"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-
 class DeepSeekV4MHCModule(Operation):
     """DeepSeek-V4 manifold-constrained hyper-connection pre/post module."""
